refactor: begin code splitting to separate files
